backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dateparser==1.2.1
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.3.2
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.2
pydantic_core==2.33.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.18.3
